+++ .idx/dev.nix
@@ -10,7 +10,13 @@
       web = {
         command = [
           "npm"
+          "run"
           "start"
+          "--"
+          "--host"
+          "0.0.0.0"
+          "--port"
+          "$PORT"
         ];
         manager = "web";
       };