+++ src/components/Tabs.tsx
@@ -43,19 +43,19 @@ const TabItem: React.FC<TabItemProps> = ({
 /**
  * Material Design 3 Tabs component
  *
- * @param props TabsProps
- * @param props.id Element id
- * @param props.items Array of TabItem configurations
- * @param props.items[].key Unique key for the item
- * @param props.items[].id Element id for the item
- * @param props.items[].label Label text
- * @param props.items[].leadingIcon Leading icon element
- * @param props.items[].active Active state
- * @param props.items[].disabled Disabled state
- * @param props.items[].onClick Click handler
- * @param props.items[].href Link URL (renders as anchor tag)
- * @param props.style Custom inline styles
- * @returns JSX.Element
+ * @param {TabsProps} props TabsProps
+ * @param {string} [props.id] Element id
+ * @param {Iterable<TabItemProps>} props.items Array of TabItem configurations
+ * @param {(string | number)} [props.items[].key] Unique key for the item
+ * @param {string} [props.items[].id] Element id for the item
+ * @param {string} props.items[].label Label text
+ * @param {React.ReactNode} [props.items[].leadingIcon] Leading icon element
+ * @param {boolean} [props.items[].active] Active state
+ * @param {boolean} [props.items[].disabled] Disabled state
+ * @param {() => void} [props.items[].onClick] Click handler
+ * @param {string} [props.items[].href] Link URL (renders as anchor tag)
+ * @param {React.CSSProperties} [props.style] Custom inline styles
+ * @returns {JSX.Element}
  *
  * @example
  * <Tabs